Crate rusoto_medialive

API for AWS Elemental MediaLive

If you're using the service, you're probably looking for MediaLiveClient and MediaLive.

Structs

AacSettings
Placeholder documentation for AacSettings

Ac3Settings
Placeholder documentation for Ac3Settings

AccessDenied
Placeholder documentation for AccessDenied

ArchiveContainerSettings
Placeholder documentation for ArchiveContainerSettings

ArchiveGroupSettings
Placeholder documentation for ArchiveGroupSettings

ArchiveOutputSettings
Placeholder documentation for ArchiveOutputSettings

AribDestinationSettings
Placeholder documentation for AribDestinationSettings

AribSourceSettings
Placeholder documentation for AribSourceSettings

AudioChannelMapping
Placeholder documentation for AudioChannelMapping

AudioCodecSettings
Placeholder documentation for AudioCodecSettings

AudioDescription
Placeholder documentation for AudioDescription

AudioLanguageSelection
Placeholder documentation for AudioLanguageSelection

AudioNormalizationSettings
Placeholder documentation for AudioNormalizationSettings

AudioOnlyHlsSettings
Placeholder documentation for AudioOnlyHlsSettings

AudioPidSelection
Placeholder documentation for AudioPidSelection

AudioSelector
Placeholder documentation for AudioSelector

AudioSelectorSettings
Placeholder documentation for AudioSelectorSettings

AvailBlanking
Placeholder documentation for AvailBlanking

AvailConfiguration
Placeholder documentation for AvailConfiguration

AvailSettings
Placeholder documentation for AvailSettings

BlackoutSlate
Placeholder documentation for BlackoutSlate

BurnInDestinationSettings
Placeholder documentation for BurnInDestinationSettings

CaptionDescription
Output groups for this Live Event. Output groups contain information about where streams should be distributed.

CaptionDestinationSettings
Placeholder documentation for CaptionDestinationSettings

CaptionLanguageMapping
Maps a caption channel to an ISO 693-2 language code (http://www.loc.gov/standards/iso639-2), with an optional description.

CaptionSelector
Output groups for this Live Event. Output groups contain information about where streams should be distributed.

CaptionSelectorSettings
Placeholder documentation for CaptionSelectorSettings

Channel
Placeholder documentation for Channel

ChannelConfigurationValidationError
Placeholder documentation for ChannelConfigurationValidationError

ChannelEgressEndpoint
Placeholder documentation for ChannelEgressEndpoint

ChannelSummary
Placeholder documentation for ChannelSummary

CreateChannel
Placeholder documentation for CreateChannel

CreateChannelRequest
A request to create a channel

CreateChannelResponse
Placeholder documentation for CreateChannelResponse

CreateChannelResultModel
Placeholder documentation for CreateChannelResultModel

CreateInput
Placeholder documentation for CreateInput

CreateInputRequest
The name of the input

CreateInputResponse
Placeholder documentation for CreateInputResponse

CreateInputResultModel
Placeholder documentation for CreateInputResultModel

CreateInputSecurityGroupRequest
The IPv4 CIDRs to whitelist for this Input Security Group

CreateInputSecurityGroupResponse
Placeholder documentation for CreateInputSecurityGroupResponse

CreateInputSecurityGroupResultModel
Placeholder documentation for CreateInputSecurityGroupResultModel

DeleteChannelRequest
Placeholder documentation for DeleteChannelRequest

DeleteChannelResponse
Placeholder documentation for DeleteChannelResponse

DeleteInputRequest
Placeholder documentation for DeleteInputRequest

DeleteInputResponse
Placeholder documentation for DeleteInputResponse

DeleteInputSecurityGroupRequest
Placeholder documentation for DeleteInputSecurityGroupRequest

DeleteInputSecurityGroupResponse
Placeholder documentation for DeleteInputSecurityGroupResponse

DeleteReservationRequest
Placeholder documentation for DeleteReservationRequest

DeleteReservationResponse
Placeholder documentation for DeleteReservationResponse

DescribeChannelRequest
Placeholder documentation for DescribeChannelRequest

DescribeChannelResponse
Placeholder documentation for DescribeChannelResponse

DescribeInputRequest
Placeholder documentation for DescribeInputRequest

DescribeInputResponse
Placeholder documentation for DescribeInputResponse

DescribeInputSecurityGroupRequest
Placeholder documentation for DescribeInputSecurityGroupRequest

DescribeInputSecurityGroupResponse
Placeholder documentation for DescribeInputSecurityGroupResponse

DescribeOfferingRequest
Placeholder documentation for DescribeOfferingRequest

DescribeOfferingResponse
Placeholder documentation for DescribeOfferingResponse

DescribeReservationRequest
Placeholder documentation for DescribeReservationRequest

DescribeReservationResponse
Placeholder documentation for DescribeReservationResponse

DvbNitSettings
DVB Network Information Table (NIT)

DvbSdtSettings
DVB Service Description Table (SDT)

DvbSubDestinationSettings
Placeholder documentation for DvbSubDestinationSettings

DvbSubSourceSettings
Placeholder documentation for DvbSubSourceSettings

DvbTdtSettings
DVB Time and Date Table (SDT)

Eac3Settings
Placeholder documentation for Eac3Settings

EmbeddedDestinationSettings
Placeholder documentation for EmbeddedDestinationSettings

EmbeddedPlusScte20DestinationSettings
Placeholder documentation for EmbeddedPlusScte20DestinationSettings

EmbeddedSourceSettings
Placeholder documentation for EmbeddedSourceSettings

Empty
Placeholder documentation for Empty

EncoderSettings
Placeholder documentation for EncoderSettings

FecOutputSettings
Placeholder documentation for FecOutputSettings

GlobalConfiguration
Placeholder documentation for GlobalConfiguration

H264Settings
Placeholder documentation for H264Settings

HlsAkamaiSettings
Placeholder documentation for HlsAkamaiSettings

HlsBasicPutSettings
Placeholder documentation for HlsBasicPutSettings

HlsCdnSettings
Placeholder documentation for HlsCdnSettings

HlsGroupSettings
Placeholder documentation for HlsGroupSettings

HlsInputSettings
Placeholder documentation for HlsInputSettings

HlsMediaStoreSettings
Placeholder documentation for HlsMediaStoreSettings

HlsOutputSettings
Placeholder documentation for HlsOutputSettings

HlsSettings
Placeholder documentation for HlsSettings

HlsWebdavSettings
Placeholder documentation for HlsWebdavSettings

Input
Placeholder documentation for Input

InputAttachment
Placeholder documentation for InputAttachment

InputChannelLevel
Placeholder documentation for InputChannelLevel

InputDestination
The settings for a PUSH type input.

InputDestinationRequest
Endpoint settings for a PUSH type input.

InputLocation
Placeholder documentation for InputLocation

InputLossBehavior
Placeholder documentation for InputLossBehavior

InputSecurityGroup
An Input Security Group

InputSecurityGroupWhitelistRequest
Request of IPv4 CIDR addresses to whitelist in a security group.

InputSettings
Live Event input parameters. There can be multiple inputs in a single Live Event.

InputSource
The settings for a PULL type input.

InputSourceRequest
Settings for for a PULL type input.

InputSpecification
Placeholder documentation for InputSpecification

InputWhitelistRule
Whitelist rule

InputWhitelistRuleCidr
An IPv4 CIDR to whitelist.

InternalServiceError
Placeholder documentation for InternalServiceError

InvalidRequest
Placeholder documentation for InvalidRequest

KeyProviderSettings
Placeholder documentation for KeyProviderSettings

LimitExceeded
Placeholder documentation for LimitExceeded

ListChannelsRequest
Placeholder documentation for ListChannelsRequest

ListChannelsResponse
Placeholder documentation for ListChannelsResponse

ListChannelsResultModel
Placeholder documentation for ListChannelsResultModel

ListInputSecurityGroupsRequest
Placeholder documentation for ListInputSecurityGroupsRequest

ListInputSecurityGroupsResponse
Placeholder documentation for ListInputSecurityGroupsResponse

ListInputSecurityGroupsResultModel
Result of input security group list request

ListInputsRequest
Placeholder documentation for ListInputsRequest

ListInputsResponse
Placeholder documentation for ListInputsResponse

ListInputsResultModel
Placeholder documentation for ListInputsResultModel

ListOfferingsRequest
Placeholder documentation for ListOfferingsRequest

ListOfferingsResponse
Placeholder documentation for ListOfferingsResponse

ListOfferingsResultModel
ListOfferings response

ListReservationsRequest
Placeholder documentation for ListReservationsRequest

ListReservationsResponse
Placeholder documentation for ListReservationsResponse

ListReservationsResultModel
ListReservations response

M2tsSettings
Placeholder documentation for M2tsSettings

M3u8Settings
Settings information for the .m3u8 container

MediaLiveClient
A client for the MediaLive API.

Mp2Settings
Placeholder documentation for Mp2Settings

MsSmoothGroupSettings
Placeholder documentation for MsSmoothGroupSettings

MsSmoothOutputSettings
Placeholder documentation for MsSmoothOutputSettings

NetworkInputSettings
Network source to transcode. Must be accessible to the Elemental Live node that is running the live event through a network connection.

Offering
Reserved resources available for purchase

Output
Output settings. There can be multiple outputs within a group.

OutputDestination
Placeholder documentation for OutputDestination

OutputDestinationSettings
Placeholder documentation for OutputDestinationSettings

OutputGroup
Output groups for this Live Event. Output groups contain information about where streams should be distributed.

OutputGroupSettings
Placeholder documentation for OutputGroupSettings

OutputLocationRef
Reference to an OutputDestination ID defined in the channel

OutputSettings
Placeholder documentation for OutputSettings

PassThroughSettings
Placeholder documentation for PassThroughSettings

PurchaseOffering
PurchaseOffering request

PurchaseOfferingRequest
Placeholder documentation for PurchaseOfferingRequest

PurchaseOfferingResponse
Placeholder documentation for PurchaseOfferingResponse

PurchaseOfferingResultModel
PurchaseOffering response

RemixSettings
Placeholder documentation for RemixSettings

Reservation
Reserved resources available to use

ReservationResourceSpecification
Resource configuration (codec, resolution, bitrate, ...)

ResourceConflict
Placeholder documentation for ResourceConflict

ResourceNotFound
Placeholder documentation for ResourceNotFound

RtmpCaptionInfoDestinationSettings
Placeholder documentation for RtmpCaptionInfoDestinationSettings

RtmpGroupSettings
Placeholder documentation for RtmpGroupSettings

RtmpOutputSettings
Placeholder documentation for RtmpOutputSettings

Scte20PlusEmbeddedDestinationSettings
Placeholder documentation for Scte20PlusEmbeddedDestinationSettings

Scte20SourceSettings
Placeholder documentation for Scte20SourceSettings

Scte27DestinationSettings
Placeholder documentation for Scte27DestinationSettings

Scte27SourceSettings
Placeholder documentation for Scte27SourceSettings

Scte35SpliceInsert
Placeholder documentation for Scte35SpliceInsert

Scte35TimeSignalApos
Placeholder documentation for Scte35TimeSignalApos

SmpteTtDestinationSettings
Placeholder documentation for SmpteTtDestinationSettings

StandardHlsSettings
Placeholder documentation for StandardHlsSettings

StartChannelRequest
Placeholder documentation for StartChannelRequest

StartChannelResponse
Placeholder documentation for StartChannelResponse

StaticKeySettings
Placeholder documentation for StaticKeySettings

StopChannelRequest
Placeholder documentation for StopChannelRequest

StopChannelResponse
Placeholder documentation for StopChannelResponse

TeletextDestinationSettings
Placeholder documentation for TeletextDestinationSettings

TeletextSourceSettings
Placeholder documentation for TeletextSourceSettings

TimecodeConfig
Placeholder documentation for TimecodeConfig

TtmlDestinationSettings
Placeholder documentation for TtmlDestinationSettings

UdpContainerSettings
Placeholder documentation for UdpContainerSettings

UdpGroupSettings
Placeholder documentation for UdpGroupSettings

UdpOutputSettings
Placeholder documentation for UdpOutputSettings

UpdateChannel
Placeholder documentation for UpdateChannel

UpdateChannelRequest
A request to update a channel.

UpdateChannelResponse
Placeholder documentation for UpdateChannelResponse

UpdateChannelResultModel
The updated channel's description.

UpdateInput
Placeholder documentation for UpdateInput

UpdateInputRequest
A request to update an input.

UpdateInputResponse
Placeholder documentation for UpdateInputResponse

UpdateInputResultModel
Placeholder documentation for UpdateInputResultModel

UpdateInputSecurityGroupRequest
The request to update some combination of the Input Security Group name and the IPv4 CIDRs the Input Security Group should allow.

UpdateInputSecurityGroupResponse
Placeholder documentation for UpdateInputSecurityGroupResponse

UpdateInputSecurityGroupResultModel
Placeholder documentation for UpdateInputSecurityGroupResultModel

ValidationError
Placeholder documentation for ValidationError

VideoCodecSettings
Placeholder documentation for VideoCodecSettings

VideoDescription
Video settings for this stream.

VideoSelector
Specifies a particular video stream within an input source. An input may have only a single video selector.

VideoSelectorPid
Placeholder documentation for VideoSelectorPid

VideoSelectorProgramId
Placeholder documentation for VideoSelectorProgramId

VideoSelectorSettings
Placeholder documentation for VideoSelectorSettings

WebvttDestinationSettings
Placeholder documentation for WebvttDestinationSettings

Enums

CreateChannelError
Errors returned by CreateChannel

CreateInputError
Errors returned by CreateInput

CreateInputSecurityGroupError
Errors returned by CreateInputSecurityGroup

DeleteChannelError
Errors returned by DeleteChannel

DeleteInputError
Errors returned by DeleteInput

DeleteInputSecurityGroupError
Errors returned by DeleteInputSecurityGroup

DeleteReservationError
Errors returned by DeleteReservation

DescribeChannelError
Errors returned by DescribeChannel

DescribeInputError
Errors returned by DescribeInput

DescribeInputSecurityGroupError
Errors returned by DescribeInputSecurityGroup

DescribeOfferingError
Errors returned by DescribeOffering

DescribeReservationError
Errors returned by DescribeReservation

ListChannelsError
Errors returned by ListChannels

ListInputSecurityGroupsError
Errors returned by ListInputSecurityGroups

ListInputsError
Errors returned by ListInputs

ListOfferingsError
Errors returned by ListOfferings

ListReservationsError
Errors returned by ListReservations

PurchaseOfferingError
Errors returned by PurchaseOffering

StartChannelError
Errors returned by StartChannel

StopChannelError
Errors returned by StopChannel

UpdateChannelError
Errors returned by UpdateChannel

UpdateInputError
Errors returned by UpdateInput

UpdateInputSecurityGroupError
Errors returned by UpdateInputSecurityGroup

Traits

MediaLive
Trait representing the capabilities of the MediaLive API. MediaLive clients implement this trait.