Struct rusoto_medialive::H264Settings

pub struct H264Settings {
    pub adaptive_quantization: Option<String>,
    pub afd_signaling: Option<String>,
    pub bitrate: Option<i64>,
    pub buf_fill_pct: Option<i64>,
    pub buf_size: Option<i64>,
    pub color_metadata: Option<String>,
    pub entropy_encoding: Option<String>,
    pub fixed_afd: Option<String>,
    pub flicker_aq: Option<String>,
    pub framerate_control: Option<String>,
    pub framerate_denominator: Option<i64>,
    pub framerate_numerator: Option<i64>,
    pub gop_b_reference: Option<String>,
    pub gop_closed_cadence: Option<i64>,
    pub gop_num_b_frames: Option<i64>,
    pub gop_size: Option<f64>,
    pub gop_size_units: Option<String>,
    pub level: Option<String>,
    pub look_ahead_rate_control: Option<String>,
    pub max_bitrate: Option<i64>,
    pub min_i_interval: Option<i64>,
    pub num_ref_frames: Option<i64>,
    pub par_control: Option<String>,
    pub par_denominator: Option<i64>,
    pub par_numerator: Option<i64>,
    pub profile: Option<String>,
    pub rate_control_mode: Option<String>,
    pub scan_type: Option<String>,
    pub scene_change_detect: Option<String>,
    pub slices: Option<i64>,
    pub softness: Option<i64>,
    pub spatial_aq: Option<String>,
    pub syntax: Option<String>,
    pub temporal_aq: Option<String>,
    pub timecode_insertion: Option<String>,
}

Placeholder documentation for H264Settings

Fields
adaptive_quantization: Option<String>

Adaptive quantization. Allows intra-frame quantizers to vary to improve visual quality.

afd_signaling: Option<String>

Indicates that AFD values will be written into the output stream. If afdSignaling is "auto", the system will try to preserve the input AFD value (in cases where multiple AFD values are valid). If set to "fixed", the AFD value will be the value configured in the fixedAfd parameter.

bitrate: Option<i64>

Average bitrate in bits/second. Required for VBR, CBR, and ABR. For MS Smooth outputs, bitrates must be unique when rounded down to the nearest multiple of 1000.

buf_fill_pct: Option<i64>

Percentage of the buffer that should initially be filled (HRD buffer model).

buf_size: Option<i64>

Size of buffer (HRD buffer model) in bits/second.

color_metadata: Option<String>

Includes colorspace metadata in the output.

entropy_encoding: Option<String>

Entropy encoding mode. Use cabac (must be in Main or High profile) or cavlc.

fixed_afd: Option<String>

Four bit AFD value to write on all frames of video in the output stream. Only valid when afdSignaling is set to 'Fixed'.

flicker_aq: Option<String>

If set to enabled, adjust quantization within each frame to reduce flicker or 'pop' on I-frames.

framerate_control: Option<String>

This field indicates how the output video frame rate is specified. If "specified" is selected then the output video frame rate is determined by framerateNumerator and framerateDenominator, else if "initializeFromSource" is selected then the output video frame rate will be set equal to the input video frame rate of the first input.

framerate_denominator: Option<i64>

Framerate denominator.

framerate_numerator: Option<i64>

Framerate numerator - framerate is a fraction, e.g. 24000 / 1001 = 23.976 fps.

gop_b_reference: Option<String>

Documentation update needed

gop_closed_cadence: Option<i64>

Frequency of closed GOPs. In streaming applications, it is recommended that this be set to 1 so a decoder joining mid-stream will receive an IDR frame as quickly as possible. Setting this value to 0 will break output segmenting.

gop_num_b_frames: Option<i64>

Number of B-frames between reference frames.

gop_size: Option<f64>

GOP size (keyframe interval) in units of either frames or seconds per gopSizeUnits. Must be greater than zero.

gop_size_units: Option<String>

Indicates if the gopSize is specified in frames or seconds. If seconds the system will convert the gopSize into a frame count at run time.

level: Option<String>

H.264 Level.

look_ahead_rate_control: Option<String>

Amount of lookahead. A value of low can decrease latency and memory usage, while high can produce better quality for certain content.

max_bitrate: Option<i64>

Maximum bitrate in bits/second (for VBR mode only).

min_i_interval: Option<i64>

Only meaningful if sceneChangeDetect is set to enabled. Enforces separation between repeated (cadence) I-frames and I-frames inserted by Scene Change Detection. If a scene change I-frame is within I-interval frames of a cadence I-frame, the GOP is shrunk and/or stretched to the scene change I-frame. GOP stretch requires enabling lookahead as well as setting I-interval. The normal cadence resumes for the next GOP. Note: Maximum GOP stretch = GOP size + Min-I-interval - 1

num_ref_frames: Option<i64>

Number of reference frames to use. The encoder may use more than requested if using B-frames and/or interlaced encoding.

par_control: Option<String>

This field indicates how the output pixel aspect ratio is specified. If "specified" is selected then the output video pixel aspect ratio is determined by parNumerator and parDenominator, else if "initializeFromSource" is selected then the output pixsel aspect ratio will be set equal to the input video pixel aspect ratio of the first input.

par_denominator: Option<i64>

Pixel Aspect Ratio denominator.

par_numerator: Option<i64>

Pixel Aspect Ratio numerator.

profile: Option<String>

H.264 Profile.

rate_control_mode: Option<String>

Rate control mode.

scan_type: Option<String>

Sets the scan type of the output to progressive or top-field-first interlaced.

scene_change_detect: Option<String>

Scene change detection. Inserts I-frames on scene changes when enabled.

slices: Option<i64>

Number of slices per picture. Must be less than or equal to the number of macroblock rows for progressive pictures, and less than or equal to half the number of macroblock rows for interlaced pictures. This field is optional; when no value is specified the encoder will choose the number of slices based on encode resolution.

softness: Option<i64>

Softness. Selects quantizer matrix, larger values reduce high-frequency content in the encoded image.

spatial_aq: Option<String>

If set to enabled, adjust quantization within each frame based on spatial variation of content complexity.

syntax: Option<String>

Produces a bitstream compliant with SMPTE RP-2027.

temporal_aq: Option<String>

If set to enabled, adjust quantization within each frame based on temporal variation of content complexity.

timecode_insertion: Option<String>

Determines how timecodes should be inserted into the video elementary stream. - 'disabled': Do not include timecodes - 'picTimingSei': Pass through picture timing SEI messages from the source specified in Timecode Config