Struct rusoto_medialive::MsSmoothGroupSettings

pub struct MsSmoothGroupSettings {
    pub acquisition_point_id: Option<String>,
    pub audio_only_timecode_control: Option<String>,
    pub certificate_mode: Option<String>,
    pub connection_retry_interval: Option<i64>,
    pub destination: OutputLocationRef,
    pub event_id: Option<String>,
    pub event_id_mode: Option<String>,
    pub event_stop_behavior: Option<String>,
    pub filecache_duration: Option<i64>,
    pub fragment_length: Option<i64>,
    pub input_loss_action: Option<String>,
    pub num_retries: Option<i64>,
    pub restart_delay: Option<i64>,
    pub segmentation_mode: Option<String>,
    pub send_delay_ms: Option<i64>,
    pub sparse_track_type: Option<String>,
    pub stream_manifest_behavior: Option<String>,
    pub timestamp_offset: Option<String>,
    pub timestamp_offset_mode: Option<String>,
}

Placeholder documentation for MsSmoothGroupSettings

Fields
acquisition_point_id: Option<String>

The value of the "Acquisition Point Identity" element used in each message placed in the sparse track. Only enabled if sparseTrackType is not "none".

audio_only_timecode_control: Option<String>

If set to passthrough for an audio-only MS Smooth output, the fragment absolute time will be set to the current timecode. This option does not write timecodes to the audio elementary stream.

certificate_mode: Option<String>

If set to verifyAuthenticity, verify the https certificate chain to a trusted Certificate Authority (CA). This will cause https outputs to self-signed certificates to fail.

connection_retry_interval: Option<i64>

Number of seconds to wait before retrying connection to the IIS server if the connection is lost. Content will be cached during this time and the cache will be be delivered to the IIS server once the connection is re-established.

destination: OutputLocationRef

Smooth Streaming publish point on an IIS server. Elemental Live acts as a "Push" encoder to IIS.

event_id: Option<String>

MS Smooth event ID to be sent to the IIS server.

Should only be specified if eventIdMode is set to useConfigured.

event_id_mode: Option<String>

Specifies whether or not to send an event ID to the IIS server. If no event ID is sent and the same Live Event is used without changing the publishing point, clients might see cached video from the previous run.

Options: - "useConfigured" - use the value provided in eventId - "useTimestamp" - generate and send an event ID based on the current timestamp - "noEventId" - do not send an event ID to the IIS server.

event_stop_behavior: Option<String>

When set to sendEos, send EOS signal to IIS server when stopping the event

filecache_duration: Option<i64>

Size in seconds of file cache for streaming outputs.

fragment_length: Option<i64>

Length of mp4 fragments to generate (in seconds). Fragment length must be compatible with GOP size and framerate.

input_loss_action: Option<String>

Parameter that control output group behavior on input loss.

num_retries: Option<i64>

Number of retry attempts.

restart_delay: Option<i64>

Number of seconds before initiating a restart due to output failure, due to exhausting the numRetries on one segment, or exceeding filecacheDuration.

segmentation_mode: Option<String>

When set to useInputSegmentation, the output segment or fragment points are set by the RAI markers from the input streams.

send_delay_ms: Option<i64>

Outputs that are "output locked" can use this delay. Assign a delay to the output that is "secondary". Do not assign a delay to the "primary" output. The delay means that the primary output will always reach the downstream system before the secondary, which helps ensure that the downstream system always uses the primary output. (If there were no delay, the downstream system might flip-flop between whichever output happens to arrive first.) If the primary fails, the downstream system will switch to the secondary output. When the primary is restarted, the downstream system will switch back to the primary (because once again it is always arriving first)

sparse_track_type: Option<String>

If set to scte35, use incoming SCTE-35 messages to generate a sparse track in this group of MS-Smooth outputs.

stream_manifest_behavior: Option<String>

When set to send, send stream manifest so publishing point doesn't start until all streams start.

timestamp_offset: Option<String>

Timestamp offset for the event. Only used if timestampOffsetMode is set to useConfiguredOffset.

timestamp_offset_mode: Option<String>

Type of timestamp date offset to use. - useEventStartDate: Use the date the event was started as the offset - useConfiguredOffset: Use an explicitly configured date as the offset