Trait rusoto_medialive::MediaLive

pub trait MediaLive {
    fn batch_update_schedule(
        &self, 
        input: BatchUpdateScheduleRequest
    ) -> RusotoFuture<BatchUpdateScheduleResponse, BatchUpdateScheduleError>;

    fn create_channel(
        &self, 
        input: CreateChannelRequest
    ) -> RusotoFuture<CreateChannelResponse, CreateChannelError>;

    fn create_input(
        &self, 
        input: CreateInputRequest
    ) -> RusotoFuture<CreateInputResponse, CreateInputError>;

    fn create_input_security_group(
        &self, 
        input: CreateInputSecurityGroupRequest
    ) -> RusotoFuture<CreateInputSecurityGroupResponse, CreateInputSecurityGroupError>;

    fn create_tags(
        &self, 
        input: CreateTagsRequest
    ) -> RusotoFuture<(), CreateTagsError>;

    fn delete_channel(
        &self, 
        input: DeleteChannelRequest
    ) -> RusotoFuture<DeleteChannelResponse, DeleteChannelError>;

    fn delete_input(
        &self, 
        input: DeleteInputRequest
    ) -> RusotoFuture<DeleteInputResponse, DeleteInputError>;

    fn delete_input_security_group(
        &self, 
        input: DeleteInputSecurityGroupRequest
    ) -> RusotoFuture<DeleteInputSecurityGroupResponse, DeleteInputSecurityGroupError>;

    fn delete_reservation(
        &self, 
        input: DeleteReservationRequest
    ) -> RusotoFuture<DeleteReservationResponse, DeleteReservationError>;

    fn delete_schedule(
        &self, 
        input: DeleteScheduleRequest
    ) -> RusotoFuture<DeleteScheduleResponse, DeleteScheduleError>;

    fn delete_tags(
        &self, 
        input: DeleteTagsRequest
    ) -> RusotoFuture<(), DeleteTagsError>;

    fn describe_channel(
        &self, 
        input: DescribeChannelRequest
    ) -> RusotoFuture<DescribeChannelResponse, DescribeChannelError>;

    fn describe_input(
        &self, 
        input: DescribeInputRequest
    ) -> RusotoFuture<DescribeInputResponse, DescribeInputError>;

    fn describe_input_security_group(
        &self, 
        input: DescribeInputSecurityGroupRequest
    ) -> RusotoFuture<DescribeInputSecurityGroupResponse, DescribeInputSecurityGroupError>;

    fn describe_offering(
        &self, 
        input: DescribeOfferingRequest
    ) -> RusotoFuture<DescribeOfferingResponse, DescribeOfferingError>;

    fn describe_reservation(
        &self, 
        input: DescribeReservationRequest
    ) -> RusotoFuture<DescribeReservationResponse, DescribeReservationError>;

    fn describe_schedule(
        &self, 
        input: DescribeScheduleRequest
    ) -> RusotoFuture<DescribeScheduleResponse, DescribeScheduleError>;

    fn list_channels(
        &self, 
        input: ListChannelsRequest
    ) -> RusotoFuture<ListChannelsResponse, ListChannelsError>;

    fn list_input_security_groups(
        &self, 
        input: ListInputSecurityGroupsRequest
    ) -> RusotoFuture<ListInputSecurityGroupsResponse, ListInputSecurityGroupsError>;

    fn list_inputs(
        &self, 
        input: ListInputsRequest
    ) -> RusotoFuture<ListInputsResponse, ListInputsError>;

    fn list_offerings(
        &self, 
        input: ListOfferingsRequest
    ) -> RusotoFuture<ListOfferingsResponse, ListOfferingsError>;

    fn list_reservations(
        &self, 
        input: ListReservationsRequest
    ) -> RusotoFuture<ListReservationsResponse, ListReservationsError>;

    fn list_tags_for_resource(
        &self, 
        input: ListTagsForResourceRequest
    ) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>;

    fn purchase_offering(
        &self, 
        input: PurchaseOfferingRequest
    ) -> RusotoFuture<PurchaseOfferingResponse, PurchaseOfferingError>;

    fn start_channel(
        &self, 
        input: StartChannelRequest
    ) -> RusotoFuture<StartChannelResponse, StartChannelError>;

    fn stop_channel(
        &self, 
        input: StopChannelRequest
    ) -> RusotoFuture<StopChannelResponse, StopChannelError>;

    fn update_channel(
        &self, 
        input: UpdateChannelRequest
    ) -> RusotoFuture<UpdateChannelResponse, UpdateChannelError>;

    fn update_channel_class(
        &self, 
        input: UpdateChannelClassRequest
    ) -> RusotoFuture<UpdateChannelClassResponse, UpdateChannelClassError>;

    fn update_input(
        &self, 
        input: UpdateInputRequest
    ) -> RusotoFuture<UpdateInputResponse, UpdateInputError>;

    fn update_input_security_group(
        &self, 
        input: UpdateInputSecurityGroupRequest
    ) -> RusotoFuture<UpdateInputSecurityGroupResponse, UpdateInputSecurityGroupError>;

    fn update_reservation(
        &self, 
        input: UpdateReservationRequest
    ) -> RusotoFuture<UpdateReservationResponse, UpdateReservationError>;
}

Trait representing the capabilities of the MediaLive API. MediaLive clients implement this trait.

Required methods

fn batch_update_schedule(
&self,
input: BatchUpdateScheduleRequest
) -> RusotoFuture<BatchUpdateScheduleResponse, BatchUpdateScheduleError>

Update a channel schedule

fn create_channel(
&self,
input: CreateChannelRequest
) -> RusotoFuture<CreateChannelResponse, CreateChannelError>

Creates a new channel

fn create_input(
&self,
input: CreateInputRequest
) -> RusotoFuture<CreateInputResponse, CreateInputError>

Create an input

fn create_input_security_group(
&self,
input: CreateInputSecurityGroupRequest
) -> RusotoFuture<CreateInputSecurityGroupResponse, CreateInputSecurityGroupError>

Creates a Input Security Group

fn create_tags(
&self,
input: CreateTagsRequest
) -> RusotoFuture<(), CreateTagsError>

Create tags for a resource

fn delete_channel(
&self,
input: DeleteChannelRequest
) -> RusotoFuture<DeleteChannelResponse, DeleteChannelError>

Starts deletion of channel. The associated outputs are also deleted.

fn delete_input(
&self,
input: DeleteInputRequest
) -> RusotoFuture<DeleteInputResponse, DeleteInputError>

Deletes the input end point

fn delete_input_security_group(
&self,
input: DeleteInputSecurityGroupRequest
) -> RusotoFuture<DeleteInputSecurityGroupResponse, DeleteInputSecurityGroupError>

Deletes an Input Security Group

fn delete_reservation(
&self,
input: DeleteReservationRequest
) -> RusotoFuture<DeleteReservationResponse, DeleteReservationError>

Delete an expired reservation.

fn delete_schedule(
&self,
input: DeleteScheduleRequest
) -> RusotoFuture<DeleteScheduleResponse, DeleteScheduleError>

Delete all schedule actions on a channel.

fn delete_tags(
&self,
input: DeleteTagsRequest
) -> RusotoFuture<(), DeleteTagsError>

Removes tags for a resource

fn describe_channel(
&self,
input: DescribeChannelRequest
) -> RusotoFuture<DescribeChannelResponse, DescribeChannelError>

Gets details about a channel

fn describe_input(
&self,
input: DescribeInputRequest
) -> RusotoFuture<DescribeInputResponse, DescribeInputError>

Produces details about an input

fn describe_input_security_group(
&self,
input: DescribeInputSecurityGroupRequest
) -> RusotoFuture<DescribeInputSecurityGroupResponse, DescribeInputSecurityGroupError>

Produces a summary of an Input Security Group

fn describe_offering(
&self,
input: DescribeOfferingRequest
) -> RusotoFuture<DescribeOfferingResponse, DescribeOfferingError>

Get details for an offering.

fn describe_reservation(
&self,
input: DescribeReservationRequest
) -> RusotoFuture<DescribeReservationResponse, DescribeReservationError>

Get details for a reservation.

fn describe_schedule(
&self,
input: DescribeScheduleRequest
) -> RusotoFuture<DescribeScheduleResponse, DescribeScheduleError>

Get a channel schedule

fn list_channels(
&self,
input: ListChannelsRequest
) -> RusotoFuture<ListChannelsResponse, ListChannelsError>

Produces list of channels that have been created

fn list_input_security_groups(
&self,
input: ListInputSecurityGroupsRequest
) -> RusotoFuture<ListInputSecurityGroupsResponse, ListInputSecurityGroupsError>

Produces a list of Input Security Groups for an account

fn list_inputs(
&self,
input: ListInputsRequest
) -> RusotoFuture<ListInputsResponse, ListInputsError>

Produces list of inputs that have been created

fn list_offerings(
&self,
input: ListOfferingsRequest
) -> RusotoFuture<ListOfferingsResponse, ListOfferingsError>

List offerings available for purchase.

fn list_reservations(
&self,
input: ListReservationsRequest
) -> RusotoFuture<ListReservationsResponse, ListReservationsError>

List purchased reservations.

fn list_tags_for_resource(
&self,
input: ListTagsForResourceRequest
) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>

Produces list of tags that have been created for a resource

fn purchase_offering(
&self,
input: PurchaseOfferingRequest
) -> RusotoFuture<PurchaseOfferingResponse, PurchaseOfferingError>

Purchase an offering and create a reservation.

fn start_channel(
&self,
input: StartChannelRequest
) -> RusotoFuture<StartChannelResponse, StartChannelError>

Starts an existing channel

fn stop_channel(
&self,
input: StopChannelRequest
) -> RusotoFuture<StopChannelResponse, StopChannelError>

Stops a running channel

fn update_channel(
&self,
input: UpdateChannelRequest
) -> RusotoFuture<UpdateChannelResponse, UpdateChannelError>

Updates a channel.

fn update_channel_class(
&self,
input: UpdateChannelClassRequest
) -> RusotoFuture<UpdateChannelClassResponse, UpdateChannelClassError>

Changes the class of the channel.

fn update_input(
&self,
input: UpdateInputRequest
) -> RusotoFuture<UpdateInputResponse, UpdateInputError>

Updates an input.

fn update_input_security_group(
&self,
input: UpdateInputSecurityGroupRequest
) -> RusotoFuture<UpdateInputSecurityGroupResponse, UpdateInputSecurityGroupError>

Update an Input Security Group's Whilelists.

fn update_reservation(
&self,
input: UpdateReservationRequest
) -> RusotoFuture<UpdateReservationResponse, UpdateReservationError>

Update reservation.

Loading content...