Crate rusoto_medialive

API for AWS Elemental MediaLive

If you're using the service, you're probably looking for MediaLiveClient and MediaLive.

Structs

AacSettings
Aac Settings

Ac3Settings
Ac3 Settings

ArchiveContainerSettings
Archive Container Settings

ArchiveGroupSettings
Archive Group Settings

ArchiveOutputSettings
Archive Output Settings

AribDestinationSettings
Arib Destination Settings

AribSourceSettings
Arib Source Settings

AudioChannelMapping
Audio Channel Mapping

AudioCodecSettings
Audio Codec Settings

AudioDescription
Audio Description

AudioLanguageSelection
Audio Language Selection

AudioNormalizationSettings
Audio Normalization Settings

AudioOnlyHlsSettings
Audio Only Hls Settings

AudioPidSelection
Audio Pid Selection

AudioSelector
Audio Selector

AudioSelectorSettings
Audio Selector Settings

AvailBlanking
Avail Blanking

AvailConfiguration
Avail Configuration

AvailSettings
Avail Settings

BatchScheduleActionCreateRequest
A list of schedule actions to create (in a request) or that have been created (in a response).

BatchScheduleActionCreateResult
List of actions that have been created in the schedule.

BatchScheduleActionDeleteRequest
A list of schedule actions to delete.

BatchScheduleActionDeleteResult
List of actions that have been deleted from the schedule.

BatchUpdateScheduleRequest
List of actions to create and list of actions to delete.

BatchUpdateScheduleResponse
Placeholder documentation for BatchUpdateScheduleResponse

BlackoutSlate
Blackout Slate

BurnInDestinationSettings
Burn In Destination Settings

CaptionDescription
Output groups for this Live Event. Output groups contain information about where streams should be distributed.

CaptionDestinationSettings
Caption Destination Settings

CaptionLanguageMapping
Maps a caption channel to an ISO 693-2 language code (http://www.loc.gov/standards/iso639-2), with an optional description.

CaptionSelector
Output groups for this Live Event. Output groups contain information about where streams should be distributed.

CaptionSelectorSettings
Caption Selector Settings

Channel
Placeholder documentation for Channel

ChannelEgressEndpoint
Placeholder documentation for ChannelEgressEndpoint

ChannelSummary
Placeholder documentation for ChannelSummary

CreateChannelRequest
A request to create a channel

CreateChannelResponse
Placeholder documentation for CreateChannelResponse

CreateInputRequest
The name of the input

CreateInputResponse
Placeholder documentation for CreateInputResponse

CreateInputSecurityGroupRequest
The IPv4 CIDRs to whitelist for this Input Security Group

CreateInputSecurityGroupResponse
Placeholder documentation for CreateInputSecurityGroupResponse

CreateTagsRequest
Placeholder documentation for CreateTagsRequest

DeleteChannelRequest
Placeholder documentation for DeleteChannelRequest

DeleteChannelResponse
Placeholder documentation for DeleteChannelResponse

DeleteInputRequest
Placeholder documentation for DeleteInputRequest

DeleteInputResponse
Placeholder documentation for DeleteInputResponse

DeleteInputSecurityGroupRequest
Placeholder documentation for DeleteInputSecurityGroupRequest

DeleteInputSecurityGroupResponse
Placeholder documentation for DeleteInputSecurityGroupResponse

DeleteReservationRequest
Placeholder documentation for DeleteReservationRequest

DeleteReservationResponse
Placeholder documentation for DeleteReservationResponse

DeleteScheduleRequest
Placeholder documentation for DeleteScheduleRequest

DeleteScheduleResponse
Placeholder documentation for DeleteScheduleResponse

DeleteTagsRequest
Placeholder documentation for DeleteTagsRequest

DescribeChannelRequest
Placeholder documentation for DescribeChannelRequest

DescribeChannelResponse
Placeholder documentation for DescribeChannelResponse

DescribeInputRequest
Placeholder documentation for DescribeInputRequest

DescribeInputResponse
Placeholder documentation for DescribeInputResponse

DescribeInputSecurityGroupRequest
Placeholder documentation for DescribeInputSecurityGroupRequest

DescribeInputSecurityGroupResponse
Placeholder documentation for DescribeInputSecurityGroupResponse

DescribeOfferingRequest
Placeholder documentation for DescribeOfferingRequest

DescribeOfferingResponse
Placeholder documentation for DescribeOfferingResponse

DescribeReservationRequest
Placeholder documentation for DescribeReservationRequest

DescribeReservationResponse
Placeholder documentation for DescribeReservationResponse

DescribeScheduleRequest
Placeholder documentation for DescribeScheduleRequest

DescribeScheduleResponse
Placeholder documentation for DescribeScheduleResponse

DvbNitSettings
DVB Network Information Table (NIT)

DvbSdtSettings
DVB Service Description Table (SDT)

DvbSubDestinationSettings
Dvb Sub Destination Settings

DvbSubSourceSettings
Dvb Sub Source Settings

DvbTdtSettings
DVB Time and Date Table (SDT)

Eac3Settings
Eac3 Settings

EmbeddedDestinationSettings
Embedded Destination Settings

EmbeddedPlusScte20DestinationSettings
Embedded Plus Scte20 Destination Settings

EmbeddedSourceSettings
Embedded Source Settings

EncoderSettings
Encoder Settings

FecOutputSettings
Fec Output Settings

FixedModeScheduleActionStartSettings
Start time for the action.

FollowModeScheduleActionStartSettings
Settings to specify if an action follows another.

FrameCaptureGroupSettings
Frame Capture Group Settings

FrameCaptureOutputSettings
Frame Capture Output Settings

FrameCaptureSettings
Frame Capture Settings

GlobalConfiguration
Global Configuration

H264Settings
H264 Settings

HlsAkamaiSettings
Hls Akamai Settings

HlsBasicPutSettings
Hls Basic Put Settings

HlsCdnSettings
Hls Cdn Settings

HlsGroupSettings
Hls Group Settings

HlsInputSettings
Hls Input Settings

HlsMediaStoreSettings
Hls Media Store Settings

HlsOutputSettings
Hls Output Settings

HlsSettings
Hls Settings

HlsTimedMetadataScheduleActionSettings
Settings for the action to emit HLS metadata

HlsWebdavSettings
Hls Webdav Settings

Input
Placeholder documentation for Input

InputAttachment
Placeholder documentation for InputAttachment

InputChannelLevel
Input Channel Level

InputDestination
The settings for a PUSH type input.

InputDestinationRequest
Endpoint settings for a PUSH type input.

InputDestinationVpc
The properties for a VPC type input destination.

InputLocation
Input Location

InputLossBehavior
Input Loss Behavior

InputSecurityGroup
An Input Security Group

InputSettings
Live Event input parameters. There can be multiple inputs in a single Live Event.

InputSource
The settings for a PULL type input.

InputSourceRequest
Settings for for a PULL type input.

InputSpecification
Placeholder documentation for InputSpecification

InputSwitchScheduleActionSettings
Settings for the action to switch an input.

InputVpcRequest
Settings for a private VPC Input. When this property is specified, the input destination addresses will be created in a VPC rather than with public Internet addresses. This property requires setting the roleArn property on Input creation. Not compatible with the inputSecurityGroups property.

InputWhitelistRule
Whitelist rule

InputWhitelistRuleCidr
An IPv4 CIDR to whitelist.

KeyProviderSettings
Key Provider Settings

ListChannelsRequest
Placeholder documentation for ListChannelsRequest

ListChannelsResponse
Placeholder documentation for ListChannelsResponse

ListInputSecurityGroupsRequest
Placeholder documentation for ListInputSecurityGroupsRequest

ListInputSecurityGroupsResponse
Placeholder documentation for ListInputSecurityGroupsResponse

ListInputsRequest
Placeholder documentation for ListInputsRequest

ListInputsResponse
Placeholder documentation for ListInputsResponse

ListOfferingsRequest
Placeholder documentation for ListOfferingsRequest

ListOfferingsResponse
Placeholder documentation for ListOfferingsResponse

ListReservationsRequest
Placeholder documentation for ListReservationsRequest

ListReservationsResponse
Placeholder documentation for ListReservationsResponse

ListTagsForResourceRequest
Placeholder documentation for ListTagsForResourceRequest

ListTagsForResourceResponse
Placeholder documentation for ListTagsForResourceResponse

M2tsSettings
M2ts Settings

M3u8Settings
Settings information for the .m3u8 container

MediaConnectFlow
The settings for a MediaConnect Flow.

MediaConnectFlowRequest
The settings for a MediaConnect Flow.

MediaLiveClient
A client for the MediaLive API.

MediaPackageGroupSettings
Media Package Group Settings

MediaPackageOutputDestinationSettings
Media Package Output Destination Settings

MediaPackageOutputSettings
Media Package Output Settings

Mp2Settings
Mp2 Settings

MsSmoothGroupSettings
Ms Smooth Group Settings

MsSmoothOutputSettings
Ms Smooth Output Settings

NetworkInputSettings
Network source to transcode. Must be accessible to the Elemental Live node that is running the live event through a network connection.

Offering
Reserved resources available for purchase

Output
Output settings. There can be multiple outputs within a group.

OutputDestination
Placeholder documentation for OutputDestination

OutputDestinationSettings
Placeholder documentation for OutputDestinationSettings

OutputGroup
Output groups for this Live Event. Output groups contain information about where streams should be distributed.

OutputGroupSettings
Output Group Settings

OutputLocationRef
Reference to an OutputDestination ID defined in the channel

OutputSettings
Output Settings

PassThroughSettings
Pass Through Settings

PauseStateScheduleActionSettings
Settings for the action to set pause state of a channel.

PipelinePauseStateSettings
Settings for pausing a pipeline.

PurchaseOfferingRequest
Placeholder documentation for PurchaseOfferingRequest

PurchaseOfferingResponse
Placeholder documentation for PurchaseOfferingResponse

RemixSettings
Remix Settings

Reservation
Reserved resources available to use

ReservationResourceSpecification
Resource configuration (codec, resolution, bitrate, ...)

RtmpCaptionInfoDestinationSettings
Rtmp Caption Info Destination Settings

RtmpGroupSettings
Rtmp Group Settings

RtmpOutputSettings
Rtmp Output Settings

ScheduleAction
Contains information on a single schedule action.

ScheduleActionSettings
Holds the settings for a single schedule action.

ScheduleActionStartSettings
Settings to specify the start time for an action.

Scte20PlusEmbeddedDestinationSettings
Scte20 Plus Embedded Destination Settings

Scte20SourceSettings
Scte20 Source Settings

Scte27DestinationSettings
Scte27 Destination Settings

Scte27SourceSettings
Scte27 Source Settings

Scte35DeliveryRestrictions
Corresponds to SCTE-35 deliverynotrestricted_flag parameter. To declare delivery restrictions, include this element and its four "restriction" flags. To declare that there are no restrictions, omit this element.

Scte35Descriptor
Holds one set of SCTE-35 Descriptor Settings.

Scte35DescriptorSettings
SCTE-35 Descriptor settings.

Scte35ReturnToNetworkScheduleActionSettings
Settings for a SCTE-35 returntonetwork message.

Scte35SegmentationDescriptor
Corresponds to SCTE-35 segmentation_descriptor.

Scte35SpliceInsert
Scte35 Splice Insert

Scte35SpliceInsertScheduleActionSettings
Settings for a SCTE-35 splice_insert message.

Scte35TimeSignalApos
Scte35 Time Signal Apos

Scte35TimeSignalScheduleActionSettings
Settings for a SCTE-35 time_signal.

SmpteTtDestinationSettings
Smpte Tt Destination Settings

StandardHlsSettings
Standard Hls Settings

StartChannelRequest
Placeholder documentation for StartChannelRequest

StartChannelResponse
Placeholder documentation for StartChannelResponse

StaticImageActivateScheduleActionSettings
Settings for the action to activate a static image.

StaticImageDeactivateScheduleActionSettings
Settings for the action to deactivate the image in a specific layer.

StaticKeySettings
Static Key Settings

StopChannelRequest
Placeholder documentation for StopChannelRequest

StopChannelResponse
Placeholder documentation for StopChannelResponse

TeletextDestinationSettings
Teletext Destination Settings

TeletextSourceSettings
Teletext Source Settings

TimecodeConfig
Timecode Config

TtmlDestinationSettings
Ttml Destination Settings

UdpContainerSettings
Udp Container Settings

UdpGroupSettings
Udp Group Settings

UdpOutputSettings
Udp Output Settings

UpdateChannelClassRequest
Channel class that the channel should be updated to.

UpdateChannelClassResponse
Placeholder documentation for UpdateChannelClassResponse

UpdateChannelRequest
A request to update a channel.

UpdateChannelResponse
Placeholder documentation for UpdateChannelResponse

UpdateInputRequest
A request to update an input.

UpdateInputResponse
Placeholder documentation for UpdateInputResponse

UpdateInputSecurityGroupRequest
The request to update some combination of the Input Security Group name and the IPv4 CIDRs the Input Security Group should allow.

UpdateInputSecurityGroupResponse
Placeholder documentation for UpdateInputSecurityGroupResponse

UpdateReservationRequest
Request to update a reservation

UpdateReservationResponse
Placeholder documentation for UpdateReservationResponse

ValidationError
Placeholder documentation for ValidationError

VideoCodecSettings
Video Codec Settings

VideoDescription
Video settings for this stream.

VideoSelector
Specifies a particular video stream within an input source. An input may have only a single video selector.

VideoSelectorPid
Video Selector Pid

VideoSelectorProgramId
Video Selector Program Id

VideoSelectorSettings
Video Selector Settings

WebvttDestinationSettings
Webvtt Destination Settings

Enums

BatchUpdateScheduleError
Errors returned by BatchUpdateSchedule

CreateChannelError
Errors returned by CreateChannel

CreateInputError
Errors returned by CreateInput

CreateInputSecurityGroupError
Errors returned by CreateInputSecurityGroup

CreateTagsError
Errors returned by CreateTags

DeleteChannelError
Errors returned by DeleteChannel

DeleteInputError
Errors returned by DeleteInput

DeleteInputSecurityGroupError
Errors returned by DeleteInputSecurityGroup

DeleteReservationError
Errors returned by DeleteReservation

DeleteScheduleError
Errors returned by DeleteSchedule

DeleteTagsError
Errors returned by DeleteTags

DescribeChannelError
Errors returned by DescribeChannel

DescribeInputError
Errors returned by DescribeInput

DescribeInputSecurityGroupError
Errors returned by DescribeInputSecurityGroup

DescribeOfferingError
Errors returned by DescribeOffering

DescribeReservationError
Errors returned by DescribeReservation

DescribeScheduleError
Errors returned by DescribeSchedule

ListChannelsError
Errors returned by ListChannels

ListInputSecurityGroupsError
Errors returned by ListInputSecurityGroups

ListInputsError
Errors returned by ListInputs

ListOfferingsError
Errors returned by ListOfferings

ListReservationsError
Errors returned by ListReservations

ListTagsForResourceError
Errors returned by ListTagsForResource

PurchaseOfferingError
Errors returned by PurchaseOffering

StartChannelError
Errors returned by StartChannel

StopChannelError
Errors returned by StopChannel

UpdateChannelClassError
Errors returned by UpdateChannelClass

UpdateChannelError
Errors returned by UpdateChannel

UpdateInputError
Errors returned by UpdateInput

UpdateInputSecurityGroupError
Errors returned by UpdateInputSecurityGroup

UpdateReservationError
Errors returned by UpdateReservation

Traits

MediaLive
Trait representing the capabilities of the MediaLive API. MediaLive clients implement this trait.