Struct rusoto_medialive::MediaLiveClient

pub struct MediaLiveClient { /* fields omitted */ }

A client for the MediaLive API.

Methods

impl MediaLiveClient

pub fn new(region: Region) -> MediaLiveClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> MediaLiveClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,